application/requirements.txt
# EA Agentic Lab - Python Dependencies

# YAML parsing for agent configs and Obsidian frontmatter
pyyaml>=6.0

# JSONPath for playbook decision logic evaluation
jsonpath-ng>=1.6.0

# Testing
pytest>=7.4.0

# Streamlit UI
streamlit>=1.28.0

# Future dependencies (not yet used)
# anthropic>=0.18.0  # Claude API
# python-dotenv>=1.0.0  # Environment variables
# jinja2>=3.1.0  # Prompt templates
